requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
openai==0.28.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
